requirements.txt
typing-extensions==3.7.4.2
cryptography==2.9.2
pyjwt==1.7.1
